fix error and timout scenarios
